# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VehicleRepairRequest": {
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Amount",
            "type": "integer"
          },
          "inventory_id": {
            "format": "uuid",
            "title": "Inventory Id",
            "type": "string"
          }
        },
        "required": [
          "inventory_id",
          "amount"
        ],
        "title": "VehicleRepairRequest",
        "type": "object"
      },
      "VehicleResponse": {
        "properties": {
          "chassis_code": {
            "title": "Chassis Code",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "durability": {
            "title": "Durability",
            "type": "integer"
          },
          "fuel": {
            "title": "Fuel",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_id",
          "code",
          "chassis_code",
          "durability",
          "fuel",
          "state",
          "version"
        ],
        "title": "VehicleResponse",
        "type": "object"
      },
...
      "app__api__schemas__JobResponse": {
        "properties": {
          "completes_at": {
            "format": "date-time",
            "title": "Completes At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_id",
          "job_type",
          "started_at",
          "completes_at",
          "state",
          "version"
        ],
        "title": "JobResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MadWorld API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/vehicles/{vehicle_id}": {
      "get": {
        "operationId": "get_vehicle_api_v1_vehicles__vehicle_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vehicle Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vehicle",
        "tags": [
          "commands"
        ]
      }
    },
...
    "/api/v1/vehicles/{vehicle_id}/repair-job": {
      "post": {
        "operationId": "start_vehicle_repair_api_v1_vehicles__vehicle_id__repair_job_post",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vehicle Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRepairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__schemas__JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Vehicle Repair",
        "tags": [
          "vehicle-repair"
        ]
      }
    },
...
  }
}